Move `getRemoteConfig` to `config/file.ts`
